Move twitter data
