+++ lib/components/timeline/common/bloc/timeline_bloc.dart
@@ -5,7 +5,7 @@ import 'package:dart_twitter_api/twitter_api.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:harpy/components/timeline/common/bloc/timeline_event.dart';
 import 'package:harpy/components/timeline/common/bloc/timeline_state.dart';
-import 'package:harpy/core/api/tweet_data.dart';
+import 'package:harpy/core/api/twitter/tweet_data.dart';
 import 'package:harpy/core/service_locator.dart';
 
 abstract class TimelineBloc extends Bloc<TimelineEvent, TimelineState> {
