+++ lib/components/tweet/widgets/media/tweet_gif.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:harpy/components/common/video_player/harpy_gif_player.dart';
 import 'package:harpy/components/common/video_player/harpy_video_player.dart';
-import 'package:harpy/core/api/media_data.dart';
+import 'package:harpy/core/api/twitter/media_data.dart';
 
 /// Builds a [HarpyVideoPlayer] for the [TweetMedia] video.
 class TweetGif extends StatelessWidget {
