+++ lib/components/tweet/widgets/tweet/content/replies.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:harpy/components/tweet/widgets/tweet/tweet_tile.dart';
-import 'package:harpy/core/api/tweet_data.dart';
+import 'package:harpy/core/api/twitter/tweet_data.dart';
 
 /// Builds [TweetTile]s for the replies in [tweet].
 ///
