+++ lib/components/tweet/widgets/tweet/tweet_tile.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:harpy/components/tweet/widgets/tweet/content/replies.dart';
 import 'package:harpy/components/tweet/widgets/tweet/content/tweet_tile_content.dart';
 import 'package:harpy/components/tweet/widgets/tweet/tweet_tile_animation.dart';
-import 'package:harpy/core/api/tweet_data.dart';
+import 'package:harpy/core/api/twitter/tweet_data.dart';
 
 /// Builds a [Card] with the [TweetTileContent] that animates when scrolling
 /// down with a [TweetTileAnimation].
